feat(java): add base route retrieval to JavaTestDataBuilder
This commit adds a new method to the JavaTestDataBuilder class called `baseRoute`. This method retrieves the base route of a given Kotlin language method. It checks if the method's containing class has the annotation "@RequestMapping" from the Spring Framework. If the annotation is found, the method retrieves the value attribute of the annotation and returns