feat(kotlin): add getTypeText() function to KotlinContextCollector
Add a new private function `getTypeText()` to the `KotlinContextCollector` class in the `KotlinContextCollector.kt` file. This function returns the type text of a given `KtTypeReference` object. The function checks if the `typeElement` is a `KtUserType` and if so, it retrieves the name of the referenced `KtClass`. If the `typeElement` is not a `KtUserType`, the function returns the text of the `KtTypeReference`.