feat(statusbar): add status event producer
This commit adds a new class called `StatusEventProducer` to the `cc.unitmesh.devti.actions.vcs` package. The `StatusEventProducer` class is responsible for triggering events related to the status of the application. It has a `triggerEvent` method that takes an `AutoDevStatus` parameter and notifies the registered listener about the event occurrence. Additionally, a new interface called `OnEventListener` is introduced, which defines the `onEventOccurred` method that will be implemented by the listener. This change is necessary to provide real-time updates to the status bar widget based on the current status of the application.