refactor(test): add line comments to related classes and current class
Add line comments to the related classes and current class in the TestCodeGenTask.kt file. This is done by retrieving the line comment prefix for the file's language and using it to format the comments. The related classes are preceded by the comment "// here are related classes:", and the current class is preceded by the comment "// here is current class information:". The comments are added to the prompter string.