feat(goland): add Go language support
Add Go language support to the Goland project by modifying the build.gradle.kts file and the go.xml resource file. The Go language API is now included as a dependency and the project type is set to "GO". This enables the use of Go language features and plugins in the Goland IDE.