chore(build): add support for multiple Go plugin versions
This commit modifies the `gradle.properties` and `build.gradle.kts` files to add support for multiple Go plugin versions. The `goPlugin233` and `goPlugin222` properties are added to the `gradle.properties` file, and the `plugins` configuration in the `build.gradle.kts` file is updated to use the appropriate Go plugin version based on the `platformVersion` property. This change allows for flexibility in selecting the desired Go plugin version for the build.