refactor(rust): rename RustContextProvider to RustVersionContextProvider
This commit renames the file `RustContextProvider.kt` to `RustVersionContextProvider.kt` in the `rust/src/main/kotlin/cc/unitmesh/rust/provider/` directory. The class `RustContextProvider` is also renamed to `RustVersionContextProvider`. The `isApplicable` method now checks if the source file language is `RsLanguage`. The `collect` method retrieves the Rust version using `CargoProjectsService` and returns a `ChatContextItem` with the Rust version information. The corresponding XML configuration file is also modified to use `RustVersionContextProvider` instead of `RustContextProvider`.