feat(context): add CppVariableContextBuilder
This commit adds a new file `CppVariableContextBuilder.kt` which implements the `VariableContextBuilder` interface. It provides the functionality to retrieve the variable context for a given `PsiElement` in C++ code. The `getVariableContext` function filters out `OCDeclaration` elements, extracts the declarators, and maps them to `VariableContext` objects. The first `VariableContext` object is returned as the result.