feat(java): add detectLanguageLevel function
Add a new function `detectLanguageLevel` to the `JavaBuildSystemProvider.kt` file. This function is responsible for detecting the language level of the Java SDK being used in the project. It checks the project SDK and module SDK to determine the language level. If the language level is detected, it returns the corresponding `LanguageLevel` object. This function will be used in the `JavaVersionProvider.kt` file to provide information about the Java SDK version being used in the project.