refactor(harmonyos): rename files to auto package
This commit renames several files in the actions package to the auto package in order to organize the codebase more effectively. The affected files are `AutoArkUiFlow.kt`, `ArkUiContext.kt`, `LayoutType.kt`, `AutoPageTask.kt`, `AutoArkUiAction.kt`, `ArkUiExample.kt`, and `ComponentType.kt`. This change improves the structure and maintainability of the code.