feat(custom-agent): refactor and add stream response action #51
The ResponseAction enumeration has been refactored to improve readability and maintainability. A new response action type, 'Stream', has been added to support streaming responses. This change also includes the removal of the 'WebView' response action, as it was not being used. The corresponding changes have been made in the `CustomAgentChatProcessor` class to handle the new 'Stream' response action.