docs(custom-agent): add PlainTextResponse to mock_market endpoint #51
This commit adds PlainTextResponse as the response class to the `mock_market` endpoint in the `server.py` file. This allows the endpoint to return plain text responses instead of JSON or HTML. Additionally, it updates the return message of the endpoint to include a list of available APIs that can be accessed.