docs(docs): update navigation order and add version label for DevIn Input Language documentation
The navigation order of the DevIn Input Language documentation has been adjusted from 1 to 4 to better reflect the structure of the development documentation. Additionally, a version label `AutoDev@1.7.1` has been added to the header, indicating the documentation's relevance to version 1.7.1 of AutoDev. This commit ensures that the documentation is better organized and more clearly associated with the correct version of the software.