fix(completion): improve performance by using ReadAction and runBlockingCancellable #101
Previously, the completion provider was using a background task to load git commits, which could slow down the completion process. This commit optimizes the performance by using ReadAction to perform the commit loading operation in a non-blocking way, and runBlockingCancellable to ensure that the operation can be cancelled if necessary. This change should result in faster and more responsive completion suggestions.