feat(exec): add CommitInsCommand for executing commits #101
Add a new class `CommitInsCommand` to the `cc.unitmesh.devti.language.compiler.exec` package. This class is responsible for executing commits to the local Git repository. It takes a `Project` object and a commit message as parameters. The `execute()` method iterates over all the change lists in the project, creates a `ChangeListCommitState` object for each change list with the list of changes and the commit message, and then runs the commit using a `LocalChangesCommitter` object. Finally, it returns "Committing..." as the result of the execution.