refactor(devins-lang): extract support for creating and running run configurations #100
This commit introduces the ability to create and run run configurations for a given project and virtual file in the `RunService` class. The `createRunSettings` function has been added to create a new run configuration settings or retrieve an existing one with the same name. The `runFile` function has been updated to use the `ExecutionManager` to execute the created run configuration. This feature enhances the development experience by providing a streamlined way to run files within the project.