refactor(devins-lang): refactor reorg conversation #100
Refactored `MessageCodeBlockCharProcessor` to improve readability and maintainability by moving constants to instance fields and adding additional checks for context changes.

In `DevInsProcessProcessor`, removed the `createTempTaskFile` function as it was not being used and replaced it with a TODO comment to indicate that functionality needs to be implemented in the future.