feat(devins-lang): add support for browsing web content with new command `/browse`
The `/browse` command has been added to the DevIns language, allowing users to fetch the content of a given URL. This feature refactors the existing command parsing logic and introduces a new `BrowseInsCommand` class to handle the execution of the `/browse` command. The command is now recognized by the DevIns parser and can be used within the language.