fix(refactor): refactor rename lookup manager listener to use custom rename lookup element #129
The rename lookup manager listener has been refactored to use a custom rename lookup element class, which improves the handling of insertions and the rendering of lookup elements with the correct icon. This change also includes the creation of a new file for the custom rename lookup element class, which is now located in the `cc.unitmesh.devti.practise` package.