feat(JavaTypeUtil): make resolveByType method public and limit class resolution
The resolveByType method in JavaTypeUtil has been made public to allow for wider usage. In JavaAutoTestService, a limit has been set on the number of classes that can be resolved at once to improve performance. If the number of resolved classes is less than or equal to the set limit, or if the element is a PsiMethod, all second children classes are loaded.