feat(RefactoringTool): enhance rename method to support method renaming #181
Updated the rename method in the JavaRefactoringTool class to support renaming of methods. This was achieved by adding a search scope to find the method in all open projects, and then applying the rename operation. Also, the rename method signature in the RefactoringTool interface was updated to take sourceName and targetName as parameters.