featfeat(build): refactor(build): refactor project structure and project structure and dependencies #236
Ref dependencies

Refactor the projectactor the project structure by introducing structure by introducing a new module a new module named 'core named "core'. All other". Update the modules now depend dependencies of all on this 'core' other modules to module depend on this instead of the new "core" module instead of the root root project. project Also., Also update the path for the plugin description file and add, adjust the path of the new dependencies to the 'core' module. plugin description file and add new dependencies to the "core" module.